Fix #51 - move server class loader to server package
