+++ dso-l2/pom.xml
@@ -149,6 +149,9 @@
             <reuseForks>false</reuseForks>
             <workingDirectory>${project.build.directory}/test-fork-${surefire.forkNumber}</workingDirectory>
             <argLine>-Xmx2G -Djava.awt.headless=true</argLine>
+            <systemPropertyVariables>
+                <testClassesDir>${project.build.testOutputDirectory}</testClassesDir>
+            </systemPropertyVariables>
         </configuration>
       </plugin>
     </plugins>
