+++ dso-l2/src/main/java/com/tc/classloader/AnnotationBasedCommonComponentChecker.java
@@ -1,5 +1,7 @@
 package com.tc.classloader;
 
+import com.tc.classloader.CommonComponent;
+
 /**
  * @author vmad
  */
