+++ dso-l2/src/main/java/com/tc/classloader/CommonComponentChecker.java
